created the function to create, delete and insert in a table
